API overhaul
- Significant modifications on the item side.  I've considered something
like this for a while, and the 1.7-related warnings from FML prompted me
to pull the trigger.  Item configurations now return "ItemDefinition"
objects for a given tag (will probably always be int), and there's now
an ItemRegistry that maps ItemDefinitions to/from actual Items, to avoid
configuration-level stuff from having to know about the actual Items.
- Recipes now use ItemDefinitions on both the input and the output ends.
This allows us to configure recipes before the actual Items are
registered, so now we don't have to worry about sequencing until we
actually get to the "register" step, where we need to make sure items
get registered before the recipes and events that use them.
- Adding a base class for CustomConfiguration that returns empty
configurations, since not all mods will need to register items, recipes,
*and* events.