More API tweaks
- ItemDefinition's Tag is now int.  I'm the boss, and I'm never going to
use a non-int tag, so there it is.
- FmlItemRegistry now allows for fetching base-game items by their
definition.  This requires mods to create ItemDefinitions for them,
which is a good thing not a bad thing.
- Making Item and Recipe registration more symmetrical.
--- ItemConfiguration requires a new method that gets the tags of new
items.
--- Added ItemRegistrar, an interface that can take an ItemConfiguration
and an ItemRegistry, and register all the configured new items using the
registry.
--- Modified RecipeRegistrar's interface method to also take in the
ItemRegistry for fetching items (instead of making it a c'tor-injected
dependency of the concrete implementation).
--- Added ItemRegistrarBase, a non-final implementation of ItemRegistrar
that only requires subclasses to be able to create Items from
ItemDefinitions.  Make sure to override CreateItemCore if you register
any new items.
--- Made CustomConfigurationBase concrete, since it's possible that mods
won't need to fetch anything from a file.

This probably looks like a lot of redundancy and excessive indirection,
but I think it really simplifies what I'd have to set up for a new mod
without relying on global state or API stability of components I don't
control.  I think externally-changing APIs will only require me to
change airbreathercore, and possibly (mitigated this as much as I could,
within reason) the ItemRegistrarBase subclasses.  Mods now just need to
do the following:
- Override CustomConfigurationBase to add support for the specific
configurations that the mod will use.
- Override ItemRegistrarBase to add support for creating the specific
new Items that the mod will use from their ItemDefinitions (or use
ItemRegistrarBase directly if there are no new items).
- Wire everything together.

In fact, even the logic for wiring everything together is ALMOST to the
point where I can make a single class in airbreathercore that just
accepts all the dependencies and goes off on its own -- only needing the
individual mods to provide the @Mod annotation and get the ball rolling.